% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRIPLEX_SIF.R
\name{TRIPLEX_SIF}
\alias{TRIPLEX_SIF}
\title{rTRIPLEX_SIF}
\usage{
TRIPLEX_SIF(Input_variable, Input_parameter)
}
\arguments{
\item{Input_variable}{A table as described in \code{\link{Inputpara}} containing the information about input variables.}

\item{Input_parameter}{A table as described in \code{\link{Inputvariable}} containing the information about input parameters.}
}
\value{
A list with class "result". More details on the output is \code{\link{result}}
}
\description{
rTRIPLEX_SIF
}
\examples{
library(rTRIPLEXSIF)
data("onemonth_exam")
data("Inputpara")
out<- TRIPLEX_SIF (Input_variable = onemonth_exam,
                   Input_parameter = Inputpara)
}
